backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.7
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.2
python-multipart==0.0.6
aiofiles==23.2.1
